Change file hash building algorithm
The file hash used to identify duplicates reads only the first 1000000 bytes of a file. Images with the same first part, for example, will end up with the same hash.
This update prepends the file size to the file contents prefix, decreasing the collision rate in this type of cases.